requirements.txt
aiogram==3.3.0
aiosqlite
python-dotenv
